Get everything working again after 8 months
This was horrific. The container flat out broke, due to the move to cgroups v2. Then ipv6 shenanigans broke the ipa install. Then SSL problems from crypto not reading the MacOS Keychain properly meant that I couldn't get the client up and running.

I switched out the ipa client I had previously fettled for one that seems to be being maintained that was forked off of the previous one I was using.